% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aa_framing}
\alias{aa_framing}
\title{AA Framing}
\format{
A data frame with 105 rows and 2 variables:
\describe{
\item{Framing}{Type of Frame}
\item{Attitude}{Attitude to AA}
\item{Resentment}{Resentment from AA}
}
}
\usage{
aa_framing
}
\description{
A dataset examining affirmative action framing
}
\keyword{datasets}
