% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ancova1}
\alias{ancova1}
\title{A dataset examining grade as a function of condition controlling for GPA}
\format{
A data frame with 72 rows and 3 variables
\describe{
\item{grade}{grade}
\item{course}{course}
\item{gpa}{gpa}
}
}
\usage{
ancova1
}
\description{
A dataset examining grade as a function of condition controlling for GPA
}
\keyword{datasets}
