% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Gaither}
\alias{Gaither}
\title{A dataset examining identification with Hillary Clinton before and after the 2016 election}
\format{
A data frame with 382 rows and 4 variables
\describe{
\item{Gender}{gender}
\item{pre}{pre}
\item{post}{post}
}
}
\usage{
Gaither
}
\description{
A dataset examining identification with Hillary Clinton before and after the 2016 election
}
\keyword{datasets}
