% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Gaitherlong}
\alias{Gaitherlong}
\title{A dataset examining identification with Hillary Clinton before and after the 2016 election}
\format{
A data frame with 382 rows and 4 variables
\describe{
\item{gender}{Gender}
\item{id}{ID}
\item{ivw}{Within group factor}
\item{identification}{identification}
}
}
\usage{
Gaitherlong
}
\description{
A dataset examining identification with Hillary Clinton before and after the 2016 election
}
\keyword{datasets}
