% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Green1}
\alias{Green1}
\title{A dataset examining responses to an interacial interaction}
\format{
A data frame with 148 rows and 4 variables
\describe{
\item{XMeta}{Overall attitude}
\item{XPosEmo}{Positive Emotions}
\item{goal_manip}{Manipulated factor}
\item{fr_manip}{Manipulated factor}
}
}
\usage{
Green1
}
\description{
A dataset examining responses to an interacial interaction
}
\keyword{datasets}
