% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{long2F}
\alias{long2F}
\title{#' A dataset examining body image as a function of an invervention}
\format{
A data frame with 4 rows and 9 variables
\describe{
\item{t1c1}{t1c1}
\item{t1c2}{t1c2}
\item{t2c1}{t2c1}
\item{t2c2}{t2c2}
\item{t3c1}{t3c1}
\item{t3c2}{t3c2}
\item{t4c1}{t4c1}
\item{t4c2}{t4c2}
}
}
\usage{
long2F
}
\description{
#' A dataset examining body image as a function of an invervention
}
\keyword{datasets}
