% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{memory}
\alias{memory}
\title{A dataset examining body image as a function of an intvervention}
\format{
A data frame with 150 rows and 3 variables examining changes in memory
\describe{
\item{id}{id}
\item{Memory}{Memory}
\item{Condition}{Condition}
}
}
\usage{
memory
}
\description{
A dataset examining body image as a function of an intvervention
}
\keyword{datasets}
