% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metalnew}
\alias{metalnew}
\title{A dataset examining sex partners as a function of group}
\format{
A data frame with 641 rows and 2 variables
\describe{
\item{SexPartner_Life}{Sexy fun time partners}
\item{Group}{Group}
}
}
\usage{
metalnew
}
\description{
A dataset examining sex partners as a function of group
}
\keyword{datasets}
