% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{missing}
\alias{missing}
\title{COVID 1 Factor}
\format{
A data frame with 389 rows and 4 variables:
\describe{
\item{political}{political orientation}
\item{group}{Type of Intervention}
\item{Beliefs}{Beliefs about COVID}
\item{SD}{Social Distancing}
}

A data frame with 480 rows and 4 variables:
\describe{
\item{Performance}{Job performance}
\item{Satifaction}{Job satifaction}
\item{WellBeing}{Well Being}
\item{Condition}{Condition}
\item{Sat_sq}{sq}
\item{Sat_log}{lg}
\item{Sat_inv}{inv}
}
}
\usage{
missing

missing
}
\description{
A dataset examining interventions regarding COVID behaviors and beliefs

A dataset examining job performance, satisfaction, well being, and research condition
}
\keyword{datasets}
