% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{missing2}
\alias{missing2}
\title{Missing data example 2}
\format{
A data frame with 480 rows and 4 variables
\describe{
\item{Performance}{Job performance}
\item{Satifaction}{Job satifaction}
\item{WellBeing}{Well Being}
\item{Condition}{Condition}
}
}
\usage{
missing2
}
\description{
A dataset examining job performance, satisfaction, well being, and research condition
}
\keyword{datasets}
