% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ML}
\alias{ML}
\title{A dataset examining sarcasm as a function of condition controlling for Agreeableness}
\format{
A data frame with 2231 rows and 3 variables
\describe{
\item{sarcasm}{sarcasm}
\item{PowerCond}{PowerCond}
\item{Agreeableness}{Agreeableness}
}
}
\usage{
ML
}
\description{
A dataset examining sarcasm as a function of condition controlling for Agreeableness
}
\keyword{datasets}
