% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{twofactorwithin}
\alias{twofactorwithin}
\title{A dataset examining body image as a function of an invervention}
\format{
A data frame with 102 rows and 4 variables
\describe{
\item{condition}{condition}
\item{id}{id}
\item{time}{time}
\item{body_image}{Body Image}
}
}
\usage{
twofactorwithin
}
\description{
A dataset examining body image as a function of an invervention
}
\keyword{datasets}
