% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{West}
\alias{West}
\title{A dataset examining body image as a function of an invervention}
\format{
A data frame with 51 rows and 3 variables
\describe{
\item{pre}{pre}
\item{post}{post}
\item{condition}{condition}
}
}
\usage{
West
}
\description{
A dataset examining body image as a function of an invervention
}
\keyword{datasets}
